+++ phydmslib/_metadata.py
@@ -1,5 +1,5 @@
-__version__ = '1.dev2'
-__author__ = 'Jesse D. Bloom'
-__url__ = 'http://jbloom.github.io/phydms'
+__version__ = '1.2.0'
+__author__ = 'the Bloom lab (see https://github.com/jbloomlab/phydms/contributors for all contributors)'
+__url__ = 'http://jbloomlab.github.io/phydms'
 __author_email__ = 'jbloom@fredhutch.org'
 __acknowledgments__ = "The 'phydms' package is written by %s; it performs likelihood calculations using the Bio++ libraries (thanks to Julien Dutheil and Laurent Gueguen)." % __author__